revert to old approach for uploading
